implement the dependency inversion principle add Core layer
