+++ Controllers/Resources/SaveVehicleResource.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel.DataAnnotations;
-using VEGA1.Models;
 
 namespace VEGA1.Controllers.Resources
 {
