+++ Core/Models/Make.cs
@@ -5,17 +5,19 @@ using System.ComponentModel.DataAnnotations;
 using System.Collections.ObjectModel;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace VEGA1.Models
-
+namespace VEGA1.Core.Models
 {
-    [Table("Models")]
-    public class Model
-    {
+    public class Make
+    { 
         public int Id { get; set; }
         [Required]
         [StringLength(255)]
         public string Name { get; set; }
-        public Make Make { get; set; }
-        public int MakeId {get; set;}
+        public ICollection<Model> Models { get; set; }
+
+        public Make()
+        {
+            Models = new Collection<Model>();
+        }
     }
 }
