+++ Core/Models/Model.cs
@@ -5,19 +5,17 @@ using System.ComponentModel.DataAnnotations;
 using System.Collections.ObjectModel;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace VEGA1.Models
+namespace VEGA1.Core.Models
+
 {
-    public class Make
-    { 
+    [Table("Models")]
+    public class Model
+    {
         public int Id { get; set; }
         [Required]
         [StringLength(255)]
         public string Name { get; set; }
-        public ICollection<Model> Models { get; set; }
-
-        public Make()
-        {
-            Models = new Collection<Model>();
-        }
+        public Make Make { get; set; }
+        public int MakeId {get; set;}
     }
 }
